Python/Django/optimize_django/requirements.txt
autopep8==1.3.3
BaseHash==3.0.4
certifi==2017.11.5
chardet==3.0.4
Django==1.11.8
django-silk==1.0.0
djangorestframework==3.7.3
Faker==0.8.7
idna==2.6
Jinja2==2.10
MarkupSafe==1.0
pycodestyle==2.3.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.1
urllib3==1.22
